+++ example/lib/main.dart
@@ -3,6 +3,7 @@ import 'package:flutter_themez/flutter_themez.dart';
 import 'package:flutter_widgetz/flutter_widgetz.dart';
 import 'package:flutter_widgetz_example/pages/accordion_page.dart';
 import 'package:flutter_widgetz_example/pages/alert_page.dart';
+import 'package:flutter_widgetz_example/pages/auth_button_page.dart';
 import 'package:flutter_widgetz_example/pages/avatar_page.dart';
 import 'package:flutter_widgetz_example/pages/badge_page.dart';
 import 'package:flutter_widgetz_example/pages/button_page.dart';
@@ -27,6 +28,7 @@ import 'package:flutter_widgetz_example/pages/radio_page.dart';
 import 'package:flutter_widgetz_example/pages/rating_page.dart';
 import 'package:flutter_widgetz_example/pages/scaffold_page.dart';
 import 'package:flutter_widgetz_example/pages/search_bar_page.dart';
+import 'package:flutter_widgetz_example/pages/shimmer_page.dart';
 import 'package:flutter_widgetz_example/pages/slider_page.dart';
 import 'package:flutter_widgetz_example/pages/sliver_grid_page.dart';
 import 'package:flutter_widgetz_example/pages/sliver_list_page.dart';
@@ -168,6 +170,7 @@ class PageInfo {
 const List<PageInfo> _pages = <PageInfo>[
   PageInfo(Icons.account_tree, AccordionPage(), 'Accordions'),
   PageInfo(Icons.notifications, AlertPage(), 'Alerts'),
+  PageInfo(Icons.login, AuthButtonPage(), 'Auth Buttons'),
   PageInfo(Icons.person, AvatarPage(), 'Avatars'),
   PageInfo(Icons.badge, BadgePage(), 'Badges'),
   PageInfo(Icons.gamepad, ButtonPage(), 'Buttons'),
@@ -190,8 +193,9 @@ const List<PageInfo> _pages = <PageInfo>[
   PageInfo(Icons.phonelink_setup_sharp, PopupMenuPage(), 'Popup Menu'),
   PageInfo(Icons.radio, RadioPage(), 'Radios'),
   PageInfo(Icons.star, RatingPage(), 'Rating Bar'),
-  PageInfo(Icons.search, SearchBarPage(), 'Search Bar'),
   PageInfo(Icons.scale, ScaffoldPage(), 'Scaffold'),
+  PageInfo(Icons.search, SearchBarPage(), 'Search Bar'),
+  PageInfo(Icons.shield, ShimmerPage(), 'Shimmer'),
   PageInfo(Icons.tune, SliderPage(), 'Sliders'),
   PageInfo(Icons.grid_4x4, SliverGridPage(), 'Sliver Grid'),
   PageInfo(Icons.list, SliverListPage(), 'Sliver List'),